feat(system): 添加部门树功能并优化相关模型
- 新增 DeptTree 类用于部门树结构
- 修改 SysDept 类，将部门 ID 类型从 Integer 改为 Long
- 更新相关控制器和服务接口，添加部门树相关方法
- 重构部门相关 VO 类，统一包名
- 优化部门树构建逻辑，支持递归构建

+++ chuang-system/chuang-system-core/src/main/java/cn/zhangchuangla/system/model/entity/SysDept.java
@@ -18,7 +18,7 @@ public class SysDept extends BaseEntity {
      * 部门ID
      */
     @TableId(type = IdType.AUTO)
-    private Integer deptId;
+    private Long deptId;
 
     /**
      * 部门名称
@@ -28,12 +28,12 @@ public class SysDept extends BaseEntity {
     /**
      * 父部门ID
      */
-    private Integer parentId;
+    private Long parentId;
 
     /**
      * 部门负责人
      */
-    private Integer managerId;
+    private Long managerId;
 
     /**
      * 部门描述
